+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
 ### 0.2.5  - 2020-04-19
 * Fix #7: Shuffled discard pile when no card left
+* Fix #25: Disable cards when it's not player's turn
 
 ### 0.2.4  - 2020-04-19
 * Better victory screen